Rename Frequency to Ingestion
